Laborator 1, Add solution for exercise 2:
For my_strlen function, I went through the elements of the array until I found
the null character and count all the characters before the null character.

For equality_check, I went through the elements of the array and searched for
the i-th element that is equal to (i + 2^i)%length element and then I print
the address.

Signed-off-by: Ceausu Matei Calin <mateicalin2002@yahoo.com>